+++ src/main/java/nextstep/app/ui/MemberController.java
@@ -4,6 +4,7 @@ import jakarta.servlet.http.HttpSession;
 import nextstep.app.domain.Member;
 import nextstep.app.domain.MemberRepository;
 import nextstep.security.context.SecurityContext;
+import nextstep.security.exceptions.AuthorizationException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
