+++ src/main/java/nextstep/security/exceptions/AuthorizationException.java
@@ -1,4 +1,4 @@
-package nextstep.app.ui;
+package nextstep.security.exceptions;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
